requirements.txt
# requirements.txt
ibapi>=9.81.1
pandas==1.5.3
numpy==1.24.3
psycopg2-binary>=2.9.1